move goal components to goals subfolder
